refactor: move IsQuorumRotationEnabled from llmq/utils to llmq/options
Also this commit add llmq/utils in modules where it is included indirectly
but it requires to be used such as rpc/quorums
