+++ src/llmq/options.cpp
@@ -3,3 +3,26 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include <llmq/options.h>
+
+#include <chainparams.h>
+#include <consensus/params.h>
+#include <deploymentstatus.h>
+
+namespace llmq
+{
+
+bool IsQuorumRotationEnabled(const Consensus::LLMQParams& llmqParams, gsl::not_null<const CBlockIndex*> pindex)
+{
+    if (!llmqParams.useRotation) {
+        return false;
+    }
+
+    int cycleQuorumBaseHeight = pindex->nHeight - (pindex->nHeight % llmqParams.dkgInterval);
+    if (cycleQuorumBaseHeight < 1) {
+        return false;
+    }
+    // It should activate at least 1 block prior to the cycle start
+    return DeploymentActiveAfter(pindex->GetAncestor(cycleQuorumBaseHeight - 1), Params().GetConsensus(), Consensus::DEPLOYMENT_DIP0024);
+}
+
+} // namespace llmq
